requirements.txt
# Core ML and NLP dependencies
torch>=2.0.0
transformers>=4.30.0
nltk>=3.8
pyphen>=0.14.0
pronouncing>=0.2.0

# Data processing and analysis
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Configuration and utilities
pydantic>=2.0.0
pyyaml>=6.0
click>=8.1.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0